src/backend/requirements.txt
# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Database
pinecone-client==3.0.0

# AI/ML
anthropic==0.8.0
openai==1.6.1
tiktoken==0.5.2

# Data Processing
numpy==1.26.2
pandas==2.1.4

# Web Scraping
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0

# Parsing
pygments==2.17.2
lark-parser==0.12.0

# Caching
redis==5.0.1
hiredis==2.3.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
mypy==1.7.1
pylint==3.0.3

# Monitoring
sentry-sdk==1.39.1
python-json-logger==2.0.7

# Utils
python-dateutil==2.8.2
pytz==2023.3.post1
